add filter label to other estimator related metrics
